+++ httpClient/writeEvents.js
@@ -17,6 +17,10 @@ module.exports = function(config) {
         return q().then(function() {
             assert(streamName, baseErr + 'Stream Name not provided');
             assert(events, baseErr + 'Events not provided');
+            assert.equal(true, events.constructor === Array, baseErr + 'Events should be an array');
+
+            if (events.length == 0)
+                return;
 
             options = options || {};
             options.expectedVersion = options.expectedVersion || -2;