+++ tcpClient/getEvents.js
@@ -13,7 +13,12 @@ module.exports = function(config) {
 
             direction = direction || 'forward';
             startPosition = startPosition == undefined && direction == 'backward' ? -1 : startPosition || 0;
-            length = length || 250;
+            length = length || 1000;
+            
+            if (length > 4096) {
+                console.warn('WARNING: Max event return limit exceeded. Using the max of 4096');
+                length = 4096;
+            }
 
             var connection = createConnection(config, reject);