+++ tests/tcp.getEvents.js
@@ -1,15 +1,15 @@
-var httpConfig = require('./support/httpConfig');
+var tcpConfig = require('./support/tcpConfig');
 var assert = require('assert');
 var eventstore = require('../index.js');
 var uuid = require('node-uuid');
 
-describe('Http Client - Get Events', function() {
+describe('TCP Client - Get Events', function() {
 
     var testStream = 'TestStream-' + uuid.v4();
     var numberOfEvents = 10;
 
     before(function() {
-        var client = eventstore.http(httpConfig);
+        var client = eventstore.tcp(tcpConfig);
 
         var events = [];
 
@@ -23,7 +23,7 @@ describe('Http Client - Get Events', function() {
     });
 
     it('Should get events reading forward', function() {
-        var client = eventstore.http(httpConfig);
+        var client = eventstore.tcp(tcpConfig);
 
         return client.getEvents(testStream, undefined, undefined, 'forward').then(function(events) {
             assert.equal(events.length, 10);
@@ -32,16 +32,16 @@ describe('Http Client - Get Events', function() {
     });
 
     it('Should get events reading backward', function() {
-        var client = eventstore.http(httpConfig);
+        var client = eventstore.tcp(tcpConfig);
 
-        return client.getEvents(testStream, 'head', undefined, 'backward').then(function(events) {
+        return client.getEvents(testStream, undefined, undefined, 'backward').then(function(events) {
             assert.equal(events.length, 10);
             assert.equal(events[0].data.something, 10);
         });
     });
 
     it('Should not get any events when start event is greater than the stream length', function() {
-        var client = eventstore.http(httpConfig);
+        var client = eventstore.tcp(tcpConfig);
 
         return client.getEvents(testStream, 11).then(function(events) {
             assert.equal(events.length, 0);
@@ -49,7 +49,7 @@ describe('Http Client - Get Events', function() {
     });
 
     it('Should get events reading backward from a start position', function() {
-        var client = eventstore.http(httpConfig);
+        var client = eventstore.tcp(tcpConfig);
 
         return client.getEvents(testStream, 2, undefined, 'backward').then(function(events) {
             assert.equal(events.length, 3);
@@ -58,7 +58,7 @@ describe('Http Client - Get Events', function() {
     });
 
     it('Should get events reading backward with a length greater than the stream length', function() {
-        var client = eventstore.http(httpConfig);
+        var client = eventstore.tcp(tcpConfig);
 
         return client.getEvents(testStream, undefined, 10000, 'backward').then(function(events) {
             assert.equal(events.length, 10);
@@ -67,7 +67,7 @@ describe('Http Client - Get Events', function() {
     });
 
     it('Should get events reading forward with a length greater than the stream length return a maximum of 4096', function() {
-        var client = eventstore.http(httpConfig);
+        var client = eventstore.tcp(tcpConfig);
 
         var testStream = 'TestStream-' + uuid.v4();
         var numberOfEvents = 5000;