+++ tests/tcp.stress.js
@@ -1,14 +1,16 @@
 var tcpConfig = require('./support/tcpConfig');
-var assert = require('assert');
 var eventstore = require('../index.js');
+var assert = require('assert');
 var uuid = require('node-uuid');
 
-describe('TCP Client - Get All Stream Events', function() {
-    it('Should write events and read back all stream events', function() {
+describe('TCP Client - Stress Tests', function() {
+    it('Should write and read back 100000 stream events', function() {
+        this.timeout(30 * 1000);
+
         var client = eventstore.tcp(tcpConfig);
 
         var events = [];
-        for (var k = 0; k < 1000; k++) {
+        for (var k = 0; k < 100000; k++) {
             events.push(eventstore.eventFactory.NewEvent('TestEventType', {
                 id: k
             }));
@@ -17,18 +19,20 @@ describe('TCP Client - Get All Stream Events', function() {
         var testStream = 'TestStream-' + uuid.v4();
         return client.writeEvents(testStream, events).then(function() {
             return client.getAllStreamEvents(testStream).then(function(events) {
-                assert.equal(events.length, 1000);
+                assert.equal(events.length, 100000);
                 assert.equal(events[0].data.id, 0);
-                assert.equal(events[999].data.id, 999)
+                assert.equal(events[99999].data.id, 99999);
             });
         });
     });
 
-    it('Should write events and read back all events from start event', function() {
+    it('Should write and read back 100000 stream events with max chunk size', function() {
+        this.timeout(30 * 1000);
+
         var client = eventstore.tcp(tcpConfig);
 
         var events = [];
-        for (var k = 0; k < 1000; k++) {
+        for (var k = 0; k < 100000; k++) {
             events.push(eventstore.eventFactory.NewEvent('TestEventType', {
                 id: k
             }));
@@ -36,10 +40,10 @@ describe('TCP Client - Get All Stream Events', function() {
 
         var testStream = 'TestStream-' + uuid.v4();
         return client.writeEvents(testStream, events).then(function() {
-            return client.getAllStreamEvents(testStream, 250, 500).then(function(events) {
-                assert.equal(events.length, 500);
-                assert.equal(events[0].data.id, 500);
-                assert.equal(events[499].data.id, 999)
+            return client.getAllStreamEvents(testStream, 4096).then(function(events) {
+                assert.equal(events.length, 100000);
+                assert.equal(events[0].data.id, 0);
+                assert.equal(events[99999].data.id, 99999);
             });
         });
     });