+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# 1.1.8 (2016-06-20)
+
+#### HTTP Client
+
+- Fix: writeEvents return successful if empty array given
+- Fix: any get events function will default to 4096 length if greater is requested (warning also displayed)
+- Feature: add getAllStreamEvents function
+
+#### TCP Client
+
+- Feature: added start event number on getAllStreamEvents
+- Fix: any get events function will default to 4096 length if greater is requested (warning also displayed)
+- Change: default chunkSize of reads from 250 to 1000 
+
+#### Tests
+
+- Added tests to TCP and HTTP client to check for undefined, empty array in writeEvents
+- Added stress tests to TCP and HTTP client
+
 # 1.1.7 (2016-06-08)
 
 #### HTTP Client